apps/linebot/requirements.txt
Flask 
pony
psycopg2-binary
python-dotenv 
line-bot-sdk
flask-cors